requirements.txt
azure-functions==1.0.5
WeasyPrint==50
